wip: Refactoring folder structure
