+++ .gitignore
@@ -8,7 +8,7 @@
 # Gradle build files
 /.gradle/
 /build/
-src/main/resources/docs/
+src/main/Oak.resources/docs/
 
 # MacOS custom attributes files created by Finder
 .DS_Store
