+++ src/main/java/Oak/Oak.java
@@ -1,0 +1,11 @@
+package Oak;
+
+import Oak.controller.TerminalController;
+
+public class Oak {
+    public static void main(String[] args) {
+        TerminalController controller = new TerminalController();
+
+        controller.start();
+    }
+}
