+++ src/main/java/Oak/task/model/Event.java
@@ -1,6 +1,6 @@
-package model;
+package Oak.task.model;
 
-import Utility.DateTimeUtility;
+import Oak.Utility.DateTimeUtility;
 
 import java.time.LocalDateTime;
