--- src/main/java/exceptions/InvalidCommandException.java
@@ -1,13 +1,0 @@
-package exceptions;
-
-import type.CommandEnum;
-
-import java.util.Arrays;
-
-public class InvalidCommandException extends InvalidInputException {
-    public InvalidCommandException(String invalidCommand) {
-        super("Invalid Command: " + invalidCommand + " received.\n" +
-                "Only the following commands are accepted:\n" +
-                Arrays.toString(CommandEnum.values()));
-    }
-}
