--- src/main/java/exceptions/InvalidInputException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidInputException extends Exception {
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
